replace array with vector

+++ include/head/core/presence_sensing_driver.hpp
@@ -30,7 +30,7 @@ class PresenceSensingDriver {
 
     auto get_tool(uint16_t reading) -> adc::Tool {
         adc::Tool td = {adc::ToolType::UNDEFINED, 0, 0};
-        for (int i = 0; i < adc::OT3ToolList.size(); i++)
+        for (size_t i = 0; i < adc::OT3ToolList.size(); i++)
             if ((reading < adc::OT3ToolList[i].detection_upper_bound) &&
                 (reading >= adc::OT3ToolList[i].detection_lower_bound))
                 return adc::OT3ToolList[i];
